+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/types/HostWithAuth.kt
@@ -1,5 +1,6 @@
 package com.github.jensim.megamanipulator.settings.types
 
+@FunctionalInterface
 interface HostWithAuth {
     fun getAuthHeaderValue(password: String?): String?
 }